+++ processing/src/main/java/org/apache/druid/segment/data/FixedIndexedWriter.java
@@ -39,7 +39,7 @@ import java.util.Iterator;
 /**
  * Writer for a {@link FixedIndexed}
  */
-public class FixedIndexedWriter<T> implements Serializer
+public class FixedIndexedWriter<T> implements DictionaryWriter<T>, Serializer
 {
   private static final int PAGE_SIZE = 4096;
   private final SegmentWriteOutMedium segmentWriteOutMedium;
@@ -73,11 +73,19 @@ public class FixedIndexedWriter<T> implements Serializer
     this.isSorted = isSorted;
   }
 
+  @Override
+  public boolean isSorted()
+  {
+    return isSorted;
+  }
+
+  @Override
   public void open() throws IOException
   {
     this.valuesOut = segmentWriteOutMedium.makeWriteOutBytes();
   }
 
+  @Override
   public int getCardinality()
   {
     return hasNulls ? numWritten + 1 : numWritten;
@@ -89,6 +97,7 @@ public class FixedIndexedWriter<T> implements Serializer
     return Byte.BYTES + Byte.BYTES + Integer.BYTES + valuesOut.size();
   }
 
+  @Override
   public void write(@Nullable T objectToWrite) throws IOException
   {
     if (prevObject != null && isSorted && comparator.compare(prevObject, objectToWrite) >= 0) {
@@ -140,6 +149,7 @@ public class FixedIndexedWriter<T> implements Serializer
   }
 
   @SuppressWarnings("unused")
+  @Override
   @Nullable
   public T get(int index) throws IOException
   {
