+++ processing/src/main/java/org/apache/druid/segment/nested/ScalarLongColumnSerializer.java
@@ -19,54 +19,28 @@
 
 package org.apache.druid.segment.nested;
 
-import com.google.common.base.Preconditions;
-import it.unimi.dsi.fastutil.ints.IntIterator;
-import org.apache.druid.collections.bitmap.ImmutableBitmap;
-import org.apache.druid.collections.bitmap.MutableBitmap;
 import org.apache.druid.java.util.common.ISE;
 import org.apache.druid.java.util.common.StringUtils;
 import org.apache.druid.java.util.common.io.Closer;
 import org.apache.druid.java.util.common.io.smoosh.FileSmoosher;
-import org.apache.druid.java.util.common.logger.Logger;
 import org.apache.druid.math.expr.ExprEval;
-import org.apache.druid.segment.ColumnValueSelector;
 import org.apache.druid.segment.IndexSpec;
 import org.apache.druid.segment.column.ColumnType;
 import org.apache.druid.segment.data.ColumnarLongsSerializer;
-import org.apache.druid.segment.data.CompressedVSizeColumnarIntsSerializer;
 import org.apache.druid.segment.data.CompressionFactory;
-import org.apache.druid.segment.data.CompressionStrategy;
-import org.apache.druid.segment.data.FixedIndexedIntWriter;
 import org.apache.druid.segment.data.FixedIndexedWriter;
-import org.apache.druid.segment.data.GenericIndexedWriter;
-import org.apache.druid.segment.data.SingleValueColumnarIntsSerializer;
 import org.apache.druid.segment.writeout.SegmentWriteOutMedium;
 
+import javax.annotation.Nullable;
 import java.io.IOException;
-import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
-import java.nio.channels.WritableByteChannel;
 
 /**
  * Serializer for a {@link ScalarLongColumn}
  */
-public class ScalarLongColumnSerializer extends NestedCommonFormatColumnSerializer
+public class ScalarLongColumnSerializer extends ScalarNestedCommonFormatColumnSerializer<Long>
 {
-  private static final Logger log = new Logger(ScalarLongColumnSerializer.class);
-
-  private final String name;
-  private final SegmentWriteOutMedium segmentWriteOutMedium;
-  private final IndexSpec indexSpec;
-  @SuppressWarnings("unused")
-  private final Closer closer;
-  private DictionaryIdLookup dictionaryIdLookup;
-  private FixedIndexedWriter<Long> longDictionaryWriter;
-  private FixedIndexedIntWriter intermediateValueWriter;
-  private boolean closedForWrite = false;
-  private boolean dictionarySerialized = false;
   private ColumnarLongsSerializer longsSerializer;
-  private ByteBuffer columnNameBytes = null;
-  private boolean hasNulls = false;
 
   public ScalarLongColumnSerializer(
       String name,
@@ -75,41 +49,36 @@ public class ScalarLongColumnSerializer extends NestedCommonFormatColumnSerializ
       Closer closer
   )
   {
-    this.name = name;
-    this.segmentWriteOutMedium = segmentWriteOutMedium;
-    this.indexSpec = indexSpec;
-    this.closer = closer;
-    this.dictionaryIdLookup = new DictionaryIdLookup();
+    super(name, LONG_DICTIONARY_FILE_NAME, indexSpec, segmentWriteOutMedium, closer);
   }
 
   @Override
-  public String getColumnName()
+  protected int processValue(@Nullable Object rawValue) throws IOException
   {
-    return name;
+    final ExprEval<?> eval = ExprEval.bestEffortOf(rawValue);
+
+    final long val = eval.asLong();
+    final int dictId = eval.isNumericNull() ? 0 : dictionaryIdLookup.lookupLong(val);
+    longsSerializer.add(dictId == 0 ? 0L : val);
+    return dictId;
   }
 
   @Override
-  public DictionaryIdLookup getGlobalLookup()
+  public void openDictionaryWriter() throws IOException
   {
-    return dictionaryIdLookup;
+    dictionaryWriter = new FixedIndexedWriter<>(
+        segmentWriteOutMedium,
+        ColumnType.LONG.getStrategy(),
+        ByteOrder.nativeOrder(),
+        Long.BYTES,
+        true
+    );
+    dictionaryWriter.open();
   }
 
   @Override
-  public boolean hasNulls()
+  protected void openValueColumnSerializer() throws IOException
   {
-    return hasNulls;
-  }
-
-  @Override
-  public void open() throws IOException
-  {
-    if (!dictionarySerialized) {
-      throw new IllegalStateException("Dictionary not serialized, cannot open value serializer");
-    }
-
-    intermediateValueWriter = new FixedIndexedIntWriter(segmentWriteOutMedium, false);
-    intermediateValueWriter.open();
-
     longsSerializer = CompressionFactory.getLongSerializer(
         name,
         segmentWriteOutMedium,
@@ -120,20 +89,6 @@ public class ScalarLongColumnSerializer extends NestedCommonFormatColumnSerializ
     );
     longsSerializer.open();
   }
-
-  @Override
-  public void openDictionaryWriter() throws IOException
-  {
-    longDictionaryWriter = new FixedIndexedWriter<>(
-        segmentWriteOutMedium,
-        ColumnType.LONG.getStrategy(),
-        ByteOrder.nativeOrder(),
-        Long.BYTES,
-        true
-    );
-    longDictionaryWriter.open();
-  }
-
 
   @Override
   public void serializeDictionaries(
@@ -148,119 +103,22 @@ public class ScalarLongColumnSerializer extends NestedCommonFormatColumnSerializ
     }
 
     // null is always 0
-    longDictionaryWriter.write(null);
+    dictionaryWriter.write(null);
     dictionaryIdLookup.addNumericNull();
 
     for (Long value : longs) {
       if (value == null) {
         continue;
       }
-      longDictionaryWriter.write(value);
+      dictionaryWriter.write(value);
       dictionaryIdLookup.addLong(value);
     }
     dictionarySerialized = true;
   }
 
   @Override
-  public void serialize(ColumnValueSelector<? extends StructuredData> selector) throws IOException
+  protected void writeValueColumn(FileSmoosher smoosher) throws IOException
   {
-    if (!dictionarySerialized) {
-      throw new ISE("Must serialize value dictionaries before serializing values for column [%s]", name);
-    }
-
-    final Object value = StructuredData.unwrap(selector.getObject());
-    final ExprEval<?> eval = ExprEval.bestEffortOf(value);
-
-    final long val = eval.asLong();
-    final int dictId = eval.isNumericNull() ? 0 : dictionaryIdLookup.lookupLong(val);
-    intermediateValueWriter.write(dictId);
-    longsSerializer.add(dictId == 0 ? 0L : val);
-    hasNulls = hasNulls || dictId == 0;
-  }
-
-
-  private void closeForWrite()
-  {
-    if (!closedForWrite) {
-      columnNameBytes = computeFilenameBytes();
-      closedForWrite = true;
-    }
-  }
-
-  @Override
-  public long getSerializedSize() throws IOException
-  {
-    closeForWrite();
-
-    long size = 1 + columnNameBytes.capacity();
-    // the value dictionaries, raw column, and null index are all stored in separate files
-    return size;
-  }
-
-  @Override
-  public void writeTo(
-      WritableByteChannel channel,
-      FileSmoosher smoosher
-  ) throws IOException
-  {
-    Preconditions.checkState(closedForWrite, "Not closed yet!");
-
-    // write out compressed dictionaryId int column and bitmap indexes by iterating intermediate value column
-    // the intermediate value column should be replaced someday by a cooler compressed int column writer that allows
-    // easy iteration of the values it writes out, so that we could just build the bitmap indexes here instead of
-    // doing both things
-    String filenameBase = StringUtils.format("%s.forward_dim", name);
-    final CompressionStrategy compression = indexSpec.getDimensionCompression();
-    final CompressionStrategy compressionToUse;
-    if (compression != CompressionStrategy.UNCOMPRESSED && compression != CompressionStrategy.NONE) {
-      compressionToUse = compression;
-    } else {
-      compressionToUse = CompressionStrategy.LZ4;
-    }
-    final SingleValueColumnarIntsSerializer encodedValueSerializer = CompressedVSizeColumnarIntsSerializer.create(
-        name,
-        segmentWriteOutMedium,
-        filenameBase,
-        longDictionaryWriter.getCardinality(),
-        compressionToUse
-    );
-    encodedValueSerializer.open();
-
-    final GenericIndexedWriter<ImmutableBitmap> bitmapIndexWriter = new GenericIndexedWriter<>(
-        segmentWriteOutMedium,
-        name,
-        indexSpec.getBitmapSerdeFactory().getObjectStrategy()
-    );
-    bitmapIndexWriter.open();
-    bitmapIndexWriter.setObjectsNotSorted();
-    final MutableBitmap[] bitmaps;
-    bitmaps = new MutableBitmap[longDictionaryWriter.getCardinality()];
-    for (int i = 0; i < bitmaps.length; i++) {
-      bitmaps[i] = indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeEmptyMutableBitmap();
-    }
-
-    final IntIterator rows = intermediateValueWriter.getIterator();
-    int rowCount = 0;
-    while (rows.hasNext()) {
-      final int dictId = rows.nextInt();
-      encodedValueSerializer.addValue(dictId);
-      bitmaps[dictId].add(rowCount++);
-    }
-
-    for (int i = 0; i < bitmaps.length; i++) {
-      final MutableBitmap bitmap = bitmaps[i];
-      bitmapIndexWriter.write(
-          indexSpec.getBitmapSerdeFactory().getBitmapFactory().makeImmutableBitmap(bitmap)
-      );
-      bitmaps[i] = null; // Reclaim memory
-    }
-
-    writeV0Header(channel, columnNameBytes);
-    writeInternal(smoosher, longDictionaryWriter, LONG_DICTIONARY_FILE_NAME);
-    writeInternal(smoosher, encodedValueSerializer, ENCODED_VALUE_COLUMN_FILE_NAME);
     writeInternal(smoosher, longsSerializer, LONG_VALUE_COLUMN_FILE_NAME);
-    writeInternal(smoosher, bitmapIndexWriter, BITMAP_INDEX_FILE_NAME);
-
-    log.info("Column [%s] serialized successfully.", name);
   }
 }
